backend/requirements.txt
fastapi
uvicorn
llama-cpp-python
pydantic
psutil
pytest